refactor(#652): move stats page to app router
