+++ frontend/src/App.tsx
@@ -36,7 +36,6 @@ import YearReviewRedirect from './profile/yearReview/YearReviewRedirect';
 import ScoreboardPage from './scoreboard/ScoreboardPage';
 import ClubScoreboardPage from './scoreboard/club/ClubScoreboardPage';
 import SearchPage from './scoreboard/search/SeachPage';
-import StatisticsPage from './scoreboard/statistics/StatisticsPage';
 import { TutorialProvider } from './tutorial/TutorialContext';
 
 LicenseInfo.setLicenseKey(
@@ -97,7 +96,6 @@ const router = createBrowserRouter(
 
                     <Route path='scoreboard'>
                         <Route index element={<ScoreboardPage />} />
-                        <Route path='stats' element={<StatisticsPage />} />
                         <Route path='search' element={<SearchPage />} />
                         <Route path='clubs/:id' element={<ClubScoreboardPage />} />
                         <Route path=':type' element={<ScoreboardPage />} />
