+++ frontend/src/app/(scoreboard)/scoreboard/stats/page.tsx
@@ -1,0 +1,5 @@
+import { StatisticsPage } from './StatisticsPage';
+
+export default function Page() {
+    return <StatisticsPage />;
+}
